more fixes of round function, corrected the how the function should be behaved
this has already been tested with builds. the previous round function is honestly incorrect, so I fixed it here. and yet the function sometimes goes -0, but it is better than the current incorrect function. so, -0 problem is not a huge bug and could be fixed later. the only thing is to get correct functions. previous decreases the number of decimals, the current commited function fixes the issue. the round function is not suppose to strictly decrease its decimals.